browser benchmark and add bluebird benchmark
