+++ .npmignore
@@ -6,4 +6,4 @@ test.js
 .npmignore
 .travis.yml
 
-benchmark.js
+benchmark/
