+++ benchmark/README.md
@@ -1,0 +1,22 @@
+polygoat benchmark
+==================
+
+```
+git clone https://github.com/sonnyp/polygoat.git
+cd polygoat
+npm install
+```
+
+## Node.js
+
+```
+npm run benchmark
+```
+
+## Browser
+
+```
+python2 -m SimpleHTTPServer 8000
+```
+
+and open [http://localhost:8000/benchmark/](http://localhost:8000/benchmark/) in your browser.
