+++ benchmark/index.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8"/>
+    <title>polygoat benchmark</title>
+    <script src="../node_modules/lodash/lodash.js"></script> <!-- benchmark.js dependency -->
+    <script src="../node_modules/platform/platform.js"></script> <!-- benchmark.js dependency -->
+    <script src="../node_modules/benchmark/benchmark.js"></script>
+    <script src="../node_modules/bluebird/js/browser/bluebird.js"></script>
+    <script src="../index.js"></script>
+    <script src="benchmark.js"></script>
+  </head>
+  <body>
+    Checkout the JavaScript console.
+  </body>
+</html>
