+++ package.json
@@ -12,7 +12,7 @@
     "callback"
   ],
   "scripts": {
-    "benchmark": "node benchmark.js",
+    "benchmark": "node benchmark/benchmark.js",
     "preversion": "npm test",
     "lint": "standard",
     "unit": "node test.js",
@@ -21,6 +21,8 @@
   "repository": "github:sonnyp/polygoat",
   "devDependencies": {
     "benchmark": "^2.1.0",
-    "bluebird": "^3.3.4"
+    "bluebird": "^3.3.4",
+    "lodash": "^4.11.1",
+    "platform": "^1.3.1"
   }
 }
